arrivalsのヘルパーのテストを追加
