+++ spec/rails_helper.rb
@@ -68,5 +68,6 @@ RSpec.configure do |config|
   end
 
   config.include Devise::Test::IntegrationHelpers, type: :system
+  config.include Devise::Test::IntegrationHelpers, type: :helper
   config.include WalkHelpers
 end
